Cache rate values from billing_rates table during MaxSessionTime to
avoid mysql queries for the same values

+++ library/rating_server.php
@@ -103,7 +103,7 @@ class Daemon {
     }
 
     function removePid() {
-        unlink($this->pid);
+        if (file_exists($this->pid)) unlink($this->pid);
     }
 }
